chore: move expression types into `expr` module
This is a backwards compatible change which also marks expression types
in the `hcl::structure` module as deprecated. These symbols will be
removed in a future release.

There are no changes to the public API of this crate except for the
newly added `expr` module.

Closes #100
