+++ src/structure/ser/structure.rs
@@ -1,4 +1,8 @@
-use super::{attribute::*, block::*};
+use super::{
+    AttributeSerializer, BlockSerializer, SerializeAttributeMap, SerializeAttributeSeq,
+    SerializeAttributeStruct, SerializeAttributeStructVariant, SerializeAttributeTupleVariant,
+    SerializeBlockSeq, SerializeBlockStruct,
+};
 use crate::{Error, Identifier, Result, Structure};
 use serde::ser::{self, Serialize, SerializeMap};
